KisBrushOp: Bypass the dabCache to allow having a seperate X & Y scaling.
Summary:
This gets KisPressureRatioOption working. Interestingly performance is
surprisingly good. Perhaps we don't need the cache?

Note the brush outline still doesn't take scaleX and scaleY into account and
shows a circle instead of an ellipse. Perhaps this is due to effectiveSpacing
only being passed scale (and not scaleX,Y).

Reviewers: rempt

Differential Revision: https://phabricator.kde.org/D1819